requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
deep-translator
speechrecognition
groq
pinecone
semantic-router
transformers
sentence-transformers
sentencepiece